feat: Segmentación en componentes del paso a paso udistrital/argo_documentacion#39
