+++ src/app/modules/registro-contrato/paso-contratistas/paso-contratistas.component.css
@@ -1,0 +1,29 @@
+.form-grid3 {
+    display: grid;
+    grid-template-columns: repeat(3, 1fr);
+    gap: 20px;
+    padding: 2%;
+}
+
+.btn-container {
+    display: flex;
+    justify-content: flex-end;
+    align-items: flex-end;
+    gap: 10px;
+}
+
+.centered-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.centered-content {
+    width: 100%;
+    text-align: center;
+    padding: 2%;
+}
+
+mat-form-field {
+    width: 100%;
+}
