+++ src/app/modules/registro-contrato/paso-info-general/paso-info-general.component.css
@@ -1,0 +1,13 @@
+.form-grid2 {
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 20px;
+    padding: 2%;
+}
+
+.btn-container {
+    display: flex;
+    justify-content: flex-end;
+    align-items: flex-end;
+    gap: 10px;
+}
