+++ src/app/modules/registro-contrato/paso-info-presupuestal/paso-info-presupuestal.component.html
@@ -1,0 +1,6 @@
+<form [formGroup]="fifthFormGroup">
+    <div class="btn-container">
+        <button mat-raised-button matStepperPrevious color="primary">Atrás</button>
+        <button mat-raised-button matStepperNext color="primary">Siguiente</button>
+    </div>
+</form>
