+++ src/app/modules/registro-contrato/paso-objeto/paso-objeto.component.css
@@ -1,0 +1,6 @@
+.btn-container {
+    display: flex;
+    justify-content: flex-end;
+    align-items: flex-end;
+    gap: 10px;
+}
