+++ src/app/modules/registro-contrato/paso-objeto/paso-objeto.component.ts
@@ -1,0 +1,44 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormBuilder, Validators, FormsModule, ReactiveFormsModule, FormGroup } from '@angular/forms';
+import { MatInputModule } from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatStepperModule } from '@angular/material/stepper';
+import { MatButtonModule } from '@angular/material/button';
+import { MatSelectModule } from '@angular/material/select';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatNativeDateModule } from '@angular/material/core';
+import { NgFor } from '@angular/common';
+import { MatTableModule } from '@angular/material/table';
+import { MatIconModule } from '@angular/material/icon';
+
+@Component({
+  selector: 'app-paso-objeto',
+  standalone: true,
+  templateUrl: './paso-objeto.component.html',
+  styleUrls: ['./paso-objeto.component.css'],
+  imports: [
+    CommonModule,
+    MatButtonModule,
+    MatStepperModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatSelectModule,
+    NgFor,
+    MatDatepickerModule,
+    MatNativeDateModule,
+    MatTableModule,
+    MatIconModule,
+  ],
+})
+export class PasoObjetoComponent {
+  // Paso 6
+  sixthFormGroup = this._formBuilder.group({
+    sixthCtrl: ['', Validators.required],
+  });
+
+  constructor(private _formBuilder: FormBuilder) { }
+
+}
