+++ src/app/modules/registro-contrato/paso-observaciones/paso-observaciones.component.html
@@ -1,0 +1,6 @@
+<form [formGroup]="seventhFormGroup">
+    <div class="btn-container">
+        <button mat-raised-button matStepperPrevious color="primary">Atrás</button>
+        <button mat-raised-button matStepperNext color="primary">Siguiente</button>
+    </div>
+</form>
