+++ src/app/modules/registro-contrato/paso-observaciones/paso-observaciones.component.ts
@@ -1,0 +1,43 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormBuilder, Validators, FormsModule, ReactiveFormsModule, FormGroup } from '@angular/forms';
+import { MatInputModule } from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatStepperModule } from '@angular/material/stepper';
+import { MatButtonModule } from '@angular/material/button';
+import { MatSelectModule } from '@angular/material/select';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatNativeDateModule } from '@angular/material/core';
+import { NgFor } from '@angular/common';
+import { MatTableModule } from '@angular/material/table';
+import { MatIconModule } from '@angular/material/icon';
+
+@Component({
+  selector: 'app-paso-observaciones',
+  standalone: true,
+  templateUrl: './paso-observaciones.component.html',
+  styleUrls: ['./paso-observaciones.component.css'],
+  imports: [
+    CommonModule,
+    MatButtonModule,
+    MatStepperModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatSelectModule,
+    NgFor,
+    MatDatepickerModule,
+    MatNativeDateModule,
+    MatTableModule,
+    MatIconModule,
+  ]
+})
+export class PasoObservacionesComponent {
+  seventhFormGroup = this._formBuilder.group({
+    seventhCtrl: ['', Validators.required],
+  });
+
+  constructor(private _formBuilder: FormBuilder) { }
+
+}
