+++ src/app/modules/registro-contrato/paso-supervisores/paso-supervisores.component.html
@@ -1,0 +1,120 @@
+<form [formGroup]="thirdFormGroup">
+    <h3>Solicitante</h3>
+    <div class="form-grid2">
+        <mat-form-field>
+            <mat-label>Dependencia:*</mat-label>
+            <mat-select>
+                <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+            </mat-select>
+        </mat-form-field>
+        <mat-form-field>
+            <mat-label>Sede:*</mat-label>
+            <mat-select>
+                <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+            </mat-select>
+        </mat-form-field>
+    </div>
+    <h3>Datos del Supervisor</h3>
+    <div *ngFor="let supervisor of supervisores; let i = index">
+        <h4>Supervisor {{i + 1}}</h4>
+        <div class="form-grid2">
+            <mat-form-field>
+                <mat-label>Dependencia:*</mat-label>
+                <mat-select [(ngModel)]="supervisor.dependencia">
+                    <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+                </mat-select>
+            </mat-form-field>
+            <mat-form-field>
+                <mat-label>Sede:*</mat-label>
+                <mat-select [(ngModel)]="supervisor.sede">
+                    <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+                </mat-select>
+            </mat-form-field>
+            <mat-form-field>
+                <mat-label>Supervisor:*</mat-label>
+                <input matInput [(ngModel)]="supervisor.nombre">
+            </mat-form-field>
+            <mat-form-field>
+                <mat-label>Cargo:*</mat-label>
+                <input matInput [(ngModel)]="supervisor.cargo">
+            </mat-form-field>
+            <mat-form-field>
+                <mat-label>Tipo de Control:*</mat-label>
+                <mat-select [(ngModel)]="supervisor.tipoControl">
+                    <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+                </mat-select>
+            </mat-form-field>
+            <mat-form-field>
+                <mat-label>Código de Verificación:*</mat-label>
+                <input matInput [(ngModel)]="supervisor.codigoVerificacion">
+            </mat-form-field>
+        </div>
+        <div class="btn-container">
+            <button mat-icon-button color="primary" (click)="eliminarSupervisor(i)"><mat-icon>delete</mat-icon></button>
+        </div>
+    </div>
+    <div class="btn-container">
+        <button mat-raised-button color="primary" (click)="agregarSupervisor()">Agregar Supervisor</button>
+    </div>
+    <h3>Lugar de Ejecución</h3>
+    <div class="form-grid3">
+        <mat-form-field>
+            <mat-label>Pais:*</mat-label>
+            <mat-select>
+                <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+            </mat-select>
+        </mat-form-field>
+        <mat-form-field>
+            <mat-label>Departamento:*</mat-label>
+            <mat-select>
+                <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+            </mat-select>
+        </mat-form-field>
+        <mat-form-field>
+            <mat-label>Sede:*</mat-label>
+            <mat-select>
+                <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+            </mat-select>
+        </mat-form-field>
+        <mat-form-field>
+            <mat-label>Municipio/Ciudad:*</mat-label>
+            <mat-select>
+                <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+            </mat-select>
+        </mat-form-field>
+        <mat-form-field>
+            <mat-label>Dependencia:*</mat-label>
+            <mat-select>
+                <!--<mat-option *ngFor="" [value]="value">
+                  {{viewValue}}
+                </mat-option>-->
+            </mat-select>
+        </mat-form-field>
+        <mat-form-field>
+            <mat-label>Dirección:*</mat-label>
+            <input matInput>
+        </mat-form-field>
+    </div>
+    <div class="btn-container">
+        <button mat-raised-button matStepperPrevious color="primary">Atrás</button>
+        <button mat-raised-button matStepperNext color="primary">Siguiente</button>
+    </div>
+</form>
