+++ src/app/modules/registro-contrato/paso-supervisores/paso-supervisores.component.ts
@@ -1,0 +1,70 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { FormBuilder, Validators, FormsModule, ReactiveFormsModule, FormGroup } from '@angular/forms';
+import { MatInputModule } from '@angular/material/input';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatStepperModule } from '@angular/material/stepper';
+import { MatButtonModule } from '@angular/material/button';
+import { MatSelectModule } from '@angular/material/select';
+import { MatDatepickerModule } from '@angular/material/datepicker';
+import { MatNativeDateModule } from '@angular/material/core';
+import { NgFor } from '@angular/common';
+import { MatIconModule } from '@angular/material/icon';
+
+@Component({
+  selector: 'app-paso-supervisores',
+  standalone: true,
+  templateUrl: './paso-supervisores.component.html',
+  styleUrls: ['./paso-supervisores.component.css'],
+  imports: [
+    CommonModule,
+    MatButtonModule,
+    MatStepperModule,
+    FormsModule,
+    ReactiveFormsModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatSelectModule,
+    NgFor,
+    MatDatepickerModule,
+    MatNativeDateModule,
+    MatIconModule
+  ]
+})
+
+export class PasoSupervisoresComponent {
+  thirdFormGroup = this._formBuilder.group({
+    thirdCtrl: ['', Validators.required],
+  });
+
+  supervisores: any[] = [{
+    dependencia: '',
+    sede: '',
+    nombre: '',
+    cargo: '',
+    tipoControl: '',
+    codigoVerificacion: ''
+  }];
+
+  agregarSupervisor() {
+    this.supervisores.push({
+      dependencia: '',
+      sede: '',
+      nombre: '',
+      cargo: '',
+      tipoControl: '',
+      codigoVerificacion: ''
+    });
+  }
+
+  eliminarSupervisor(index: number) {
+    if (this.supervisores.length > 1) {
+      this.supervisores.splice(index, 1);
+    } else {
+      // Opcional: mostrar un mensaje si intentan eliminar el último supervisor
+      console.log('No se puede eliminar el último supervisor');
+    }
+  }
+
+  constructor(private _formBuilder: FormBuilder) { }
+}
