+++ src/app/modules/registro-contrato/registro-contrato.component.css
@@ -21,54 +21,3 @@
     white-space: normal;
     overflow: visible;
 }
-
-.form-grid1 {
-    padding: 2%;
-    gap: 20px;
-}
-
-.form-grid2 {
-    display: grid;
-    grid-template-columns: repeat(2, 1fr);
-    gap: 20px;
-    padding: 2%;
-}
-
-.form-grid3 {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 20px;
-    padding: 2%;
-}
-
-mat-table mat-cell,
-mat-table mat-header-cell {
-    border: 1px solid gray;
-}
-
-mat-table {
-    padding-bottom: 2%;
-}
-
-.btn-container {
-    display: flex;
-    justify-content: flex-end;
-    align-items: flex-end;
-    gap: 10px;
-}
-
-.centered-container {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-}
-
-.centered-content {
-    width: 100%;
-    text-align: center;
-    padding: 2%;
-}
-
-mat-form-field {
-    width: 100%;
-}
